refactor(renderer): finish worktree palette extraction
(cherry picked from commit ea8e6e595ad66304666c87777c791792b6bc582a)
