+++ src/renderer/src/components/use-worktree-jump-palette-sections.ts
@@ -54,7 +54,9 @@ export function useWorktreeJumpPaletteSections({
   setExpandedSectionCaps
 }: WorktreeJumpPaletteSectionsInput) {
   const openTabsLeadSections = useMemo(() => {
-    if (!hasQuery) return true
+    if (!hasQuery) {
+      return true
+    }
     return shouldOpenTabsLeadPaletteSections({
       bestWorktreeQualityRank: worktreeItems[0]
         ? bestPaletteQualityRank([worktreeItems[0].match.qualityClass])
@@ -66,7 +68,9 @@ export function useWorktreeJumpPaletteSections({
   }, [hasQuery, openTabItems, worktreeItems])
 
   const middleLeadsSections = useMemo(() => {
-    if (!hasQuery) return false
+    if (!hasQuery) {
+      return false
+    }
     const bestEntityQualityRank = Math.min(
       worktreeItems[0]
         ? bestPaletteQualityRank([worktreeItems[0].match.qualityClass])
